chore(dev): require facets on demand
